refactor: move metrics package to module.
