+++ pykp/metrics.py
@@ -1,30 +1,8 @@
 """
-Provides an implementation of the phase transition of the knapsack
-problem, based on the following paper: Yadav, Nitin, et al. "Phase transition
-in the knapsack problem." arXiv preprint arXiv:1806.10244 (2018).
-
-Example:
-    Compute the phase transition of the knapsack problem::
-
-        from pykp.metrics import phase_transition
-
-        grid, solvability_matrix = phase_transition(
-            num_items=10,
-            samples=100,
-            resolution=(20, 20),
-        )
-
-    Save the phase transition to a CSV file by specifying an optional `path`
-    argument ::
-
-        phase_transition(
-            num_items=10,
-            samples=100,
-            resolution=(20, 20),
-            path="phase_transition.csv",
-        )
+This module contains various metrics for evaluating knapsack problem instances.
 """
 
+import itertools
 from typing import Callable
 
 import numpy as np
@@ -33,7 +11,8 @@ from tqdm import tqdm
 
 from pykp.item import Item
 
-from .. import solvers
+from . import solvers
+from .arrangement import Arrangement
 
 SOLVERS = ["branch_and_bound", "mzn_gecode"]
 
@@ -218,10 +197,80 @@ def phase_transition(
     return grid, phase_transition
 
 
-if __name__ == "__main__":
-    phase_transition(
-        num_items=10,
-        samples=1,
-        resolution=(10, 10),
-        path="phase_transition.csv",
-    )
+def sahni_k(
+    arrangement: Arrangement,
+    capacity: int,
+) -> int:
+    """
+    Provides an implementation of the Sahni-K metric for evaluating
+    arrangements of items in the knapsack problem.
+
+    Example:
+        To calculate the Sahni-k of the optimal solution to a knapsack problem
+        instance, first solve the instance and then call the metric on the
+        optimal arrangement::
+
+            from pyinstance import Knapsack
+            from pyinstance import Item
+            import pyinstance.metrics as metrics
+
+            items = [
+                Item(value=10, weight=5),
+                Item(value=15, weight=10),
+                Item(value=7, weight=3),
+            ]
+            capacity = 15
+            instance = Knapsack(items=items, capacity=capacity)
+            instance.solve()
+
+            sahni_k = metrics.sahni_k(instance.optimal_nodes[0], capacity)
+            print(sahni_k)
+
+    Parameters:
+        arrangement (Arrangement): The arrangement for which to calculate
+            Sahni-k.
+        capacity (int): The capacity of the knapsack.
+
+    Returns:
+        int: Sahni-k value.
+    """
+    if not isinstance(arrangement, Arrangement):
+        raise ValueError("`arrangement` must be of type `Arrangement`.")
+    if arrangement.weight > capacity:
+        raise ValueError(
+            """The total weight of items included in the `Arrangement` exceeds 
+            the `capacity`."""
+        )
+
+    in_items = [
+        arrangement.items[i]
+        for i, element in enumerate(arrangement.state)
+        if element == 1
+    ]
+    for subset_size in range(0, len(arrangement.state) + 1):
+        for subset in itertools.combinations(in_items, subset_size):
+            subset = list(subset)
+            weight = sum([item.weight for item in subset])
+
+            # Solve greedily
+            while True:
+                if len(subset) == len(arrangement.items):
+                    break
+
+                # Check instance at capacity
+                out_items = [
+                    item for item in arrangement.items if item not in subset
+                ]
+                if (
+                    min([weight + item.weight for item in out_items])
+                    > capacity
+                ):
+                    break
+
+                densities = [item.value / item.weight for item in out_items]
+                max_density_item = out_items[densities.index(max(densities))]
+                subset.append(max_density_item)
+                weight = sum([item.weight for item in subset])
+
+            if set(subset) == set(in_items):
+                return subset_size
