--- pykp/metrics/__init__.py
@@ -1,8 +1,0 @@
-"""
-This module contains various metrics for evaluating knapsack problem instances.
-"""
-
-from .phase_transition import phase_transition
-from .sahni_k import sahni_k
-
-__all__ = ["sahni_k", "phase_transition"]
